Update path to the appropriate section
It appears this has changed a bit from when the docs were written.

It took a bit of time to figure out where to look to find the mentioned
toggle. It seems like it might have been in a different place before.
Currently, you have to click the "UI Extensions" section under the
"FEATURES" heading to see the toggle. We update the documentation to
reflect that.