Remove the install step
It appears this step is automatic now.

It might be that this step still should exist, or that it's named
differently or something else. It doesn't look like there's anything to
install nowadays though.

If we do still need an install step, we should revert this change to
bring it back.