Document optional arguments when opening a modal
We want to show how you might pass through additional arguments.

This is similar to how side panels work. So much of the same
understanding applies here.

We add some custom data to the modal open call in the starter kit so
it's more similar to side panels.

We're tracking a branch for the SDK, since it needs to be
merged/released first. We'll need to update that before merging this
code into the `master` branch.