Use `unknown` instead of `any`
We want to be a bit more type safe with the args we pass through.

Since we cannot know what data we're passing around, we don't want to
type it as though you can do anything with it. The main difference
between `any` and `unknown` is that TypeScript will yell if you do
anything except cast or pass along a value of type `unknown`:
https://www.typescriptlang.org/docs/handbook/2/functions.html#unknown.

This is effectively the same as before, but an SDK user has to be
explicit with giving the `args` a type before using it.