Make `AuthState` generic
We want to remove the `any`.

We make it so the `args` of the `AuthState` is now generic and can be
decided when initializing the client. This should allow for greater type
safety when working with the SDK client and using auth in particular.

We thread around the `AuthStateArgs` to the places that seem
appropriate; namely the `DDClient`, `DDAuthClient`, and
`DDEventsClient`. Everything else gets defaulted to `unknown` (since the
`AuthState` doesn't seem to propagate to those other places). If we find
we need this type variable in other places, we can propagate it there as
well.

We also add a few tests to check that the inference is working the way
we expect it to. These tests are a bit unconventional in that they don't
actually need to run or anything. We just want `tsc` to check the types
in an automated fashion, but don't have a good way of enforcing that in
the "source" files.

While making this change, it was a bit unclear why things weren't
inferring properly when it seemed like they should've. It turned out to
be a minor typo. So we're going to try and mitigate dealing with that
again when we change this code. If we've a different way of solving the
same problem, we can convert to that instead.