Split `WidgetOptionItem` based on `type`
We want to remove the `any` and be a bit more precise.

We split things into a base, boolean, and string interfaces. This allows
us to replace the `any` with the appropriate type for each `type`. This
also means that we can remove the `enum` entirely from
`WidgetOptionItemBoolean` (since there's no valid enumeration that isn't
`false`/`true`).