Make side panel args `unknown`
We want to remove the `any`.

Although it might seem like it makes sense to use a type variable here,
it doesn't really give us any additional type safety. Since there would
be nothing for that type variable to line up with, it'd be as effective
as `unknown`.

The point where this would become useful is if we had something  else
(like on the `DDEventsClient` side) with a matching type variable. Since
that's ends up being a very large change (we have to thread around
multiple type variables everywhere) we stave off making that change for
now. Maybe in the future it will make more sense to do it, so we can
revisit it at that point.