Add support for custom widget cog menu items
We want to allow Apps to add actions to the custom widget cog menu.

We've had a few requests to support actions on the custom widget cog
menu. Since there's enough interest in it, we start fleshing out the
feature.

We follow much the same pattern as the dashboard cog menu items and
widget context menu items: allow people to make either links or actions,
and handle them accordingly. For the actions to work properly, we need a
new event that should pass along the correct data.

We're also being a little stricter with our types for the click data.
Since we know we're only working with custom widgets, we don't have to
worry about passing along every different kind of widget's data. We
whittle down the `definition` so that clients can deal with an easier
type to work with than `any`.