Show an example of using the new event
We mimic what we do for the dashboard cog menu.

This event is largely the same as the dashboard cog menu/widget context
menu events. The main difference being that we've got some stricter
types on the context. Namely, the widget definition is specifically for
custom widgets. So we can make assumptions about the structure of it
that we can't do with the other events.