Add an event for resizing the iframe
We want to allow the SDK to send an event for changes in iframe size.

The idea here is that trying to size an iframe correctly is
nigh-impossible. Instead of trying to guess or do a bunch of work with
styling, we will send an event when the iframe has been resized. This
way, it can be adjusted to the dimensions provided (within reason).

Since this is an event from the SDK, we cannot guarantee that it will be
processed. At best, we'll set the height to 80% of what was given.