Extract the resizing to a public method
We want to allow people to resize an iframe when necessary.

Turns out, listening to the `'resize'` event on the iframe's window
doesn't work for 100% of the use cases we have. In particular, if an App
changes its content (something appears/disappears), the `'resize'` event
isn't fired. The `'resize'` event still works great for that initial
load. But, it doesn't work out for anything beyond that.

We tried a few approaches to work around this. One approach was to use
`ResizeObserver`:
https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserver/ResizeObserver.
This does exactly what we want, but would require a polyfill to have
enough coverage for browsers we support.

We might be able to figure some other way to automatically resize an
iframe at some point in the future. Until then, we'll need to ask App
developers to invoke the `DDClient.resize()` method when they change the
content.