Support overriding computed dimensions
We want to allow App developers to set explicit dimensions.

For the most part, App developers shouldn't really need to explicitly
set the dimensions. But, there are times when the App developer knows
better than us what size things should be. We allow them to suggest
explicit dimensions to resize, and compute anything not passed in.