Only handle one `'resize'` event
We have to be careful not to loop indefinitely.

As Noah pointed out in review:
https://github.com/DataDog/ui-extensions-sdk/pull/55#issuecomment-926942465,
there's a potential to loop indefinitely with the previous
implementation. To work around that issue, we only handle the `'resize'`
event once.

If someone wants automatic resizing behavior, they can opt-into that by
invoking the `DDClient.resize()` method themselves (much like we were
doing before). But, since we cannot know whether it'll cause a loop, we
don't want to force everyone to deal with it.