Explain more about how modal dimensions work
We want to try and clarify the behavior.

In a effort to address some of the concerns raised:
https://github.com/DataDog/ui-extensions-sdk/pull/55#issuecomment-926942465,
we try to explain more clearly that only the modal width is adjusted by
the `size` option. We also try to explain that the initial height is set
automatically, but it can be overridden with `DDClient.resize()`.