UIAPPS-77 Move to `packages/ui-extensions-sdk`
We're only moving the stuff that's specific to this package.

We want to start using a single repo for all Apps-related stuff. In
order to do that, we're going to subsume the
https://github.com/DataDog/apps repo into this one. To make that a bit
easier, we're starting by moving everything specific to the
`@datadog/ui-extensions-sdk` repo into the `packages/ui-extensions-sdk`
directory. We also turn this repo into a monorepo by using yarn
workspaces.

We have to make a few changes to some build and CI things for the new
organization. Hopefully, this will all work the first time through.