UIAPPS-78 Fix `format`, `lint`, and `test` scripts
This at least should get CI passing again.

There are a couple parts to this. First, the `examples/sentiment`
package didn't have either a `format` or `lint` script. We add a
`format` script (but don't run it), and a `lint` script that exits with
0. The code style in the `examples/sentiment` package doesn't match up
with the rest of this repo, so it will have to be fixed before we add
an actual `lint` script that does something useful.

Another part of this is that the `lint` script for
`packages/ui-extensions-sdk` broke after pulling in the code from the
`apps` repo. It's unclear why, but `eslint-plugin-prettier` seems to
have some issue with being in a monorepo when run under yarn v1 and not
at the top-level:
https://github.com/prettier/eslint-plugin-prettier/issues/396. While we
could upgrade to yarn v3, we hold off on making that change at the
moment.

Finally, we add a few ignore files so the format/lint scripts are useful
in development (where there might be artifacts in the respective
directories).