UIAPPS-72 Setup `lerna`
In order to use `auto`, we first have to use `lerna`.

`auto` depends on having `lerna` setup and working in order to do any of
its work. We bring in and init `lerna` to use an independent versioning
scheme with `yarn`. Hopefully, this will mean that we'll be able to
actual version each package independent of the others so we don't have
to bump everything all the time.

We're also only explicitly looking in the `packages` directory. Maybe
this is not quite right and we should rely on the `"workspaces"` field
in the `package.json`, but that includes the `examples` directory and
it's not clear if we'll always remember to mark each package as
`"private": true`. In any case, we can always change it later on.