UIAPPS-72 Setup `auto` for managing releases
We want to make it painless to develop in this monorepo.

One of the non-trivial issues with having a monorepo is making releases
of the packages in said monorepo. There are many tools around that can
help with that and we could also roll our own. We take a page from the
Dataviz SDK and use `auto` to manage making releases.

The way it works is laid out in the `RELEASE.md` file. We can continue
to expand on this workflow as we grow.

We also added a Pull Request template to make the workflow just a little
easier. Otherwise, it's unlikely that people will know what to add in
order to make things work properly.