Build everything in CI
We have to do a bit more than just `lint` and `test`.

The `build` step will catch things that the others do not. In
particular, the `build` step for the `sentiment` example will fail in
CI. We make that step happen in CI so we will catch it going forward.