Fix linting failures
These were two `eslint` failures that we weren't addressing.

The issue with `const client = init()` was `no-unused-vars`. Because we
were never using the `client`, it was marked as an issue.

The issue with the `export default {…}` was
`import/no-anonymous-default-export`. Because we didn't give it a name,
it was marked as an issue. Likely, we should actually export it not as a
default given our code style in other places at Datadog.