UIAPPS-72 Use `npm config` instead of `.npmrc`
We can't leave uncommitted changes.

`lerna` doesn't seem to work when there are uncommitted changes. When
we're writing the `.npmrc` file in CI, there are uncommitted changes.

One option is to add the `.npmrc` file to `.gitignore`, but that means
we can never track the `.npmrc` file (and this file has valid uses
outside of `lerna`). If we can't get other things to work, this might be
our only recourse.

Another option is to try using `npm config`. This should set the config
in a `~/.npmrc` file (which should not show up as an untracked file).
Whether `lerna` will respect that is to be discovered. Guess we'll know
if this code ends up staying after a while.