UIAPPS-69 Start an npm init package
We want to make it easier to get start creating a Datadog App.

We're trying a different approach here. We're creating an `npm init`
package that App developers can use to get started. As mentioned in the
readme, this package should allow a workflow like:
`npm init @datadog/app` or `yarn create @datadog/app`. It should do all
the work we've been asking App developers to do manually: download
files, install dependencies, etc.

The way it currently works is that it downloads a tarball of the repo,
extracts it and installs dependencies. This approach should allow us to
use the `examples` directory to fill out more examples without requiring
everyone to become an expert in how this package works. There aren't a
whole lot of examples at the moment, but this should allow us to expand
them relatively easily.