UIAPPS-87 Setup `changesets`
We don't want to use `auto` to make normal releases.

`auto` seems to be primarily focused on versioning all packages with the
same version. It has good support for that process, but not so great
support for versioning packages independently. Since we have packages
that we want to have at different versions, it's not really going to
work out for us.

`changesets` seems to be primarily focused on allowing packages to be
versioned independently. The idea is that when creating a PR, we can
throw together a description of the change–a changset–and when
releasing, it'll build that up into the actual changelog. Importantly,
each changeset can describe how the packages should change their version
and they get combined in the "natural" way: highest kind of change
(major, minor, or patch) for each package wins, but different packages
don't affect others.

Here we are only setting up the CLI and config. We'll move the process
over in a future commit.