UIAPPS-87 Move release process to `changesets`
We still keep `auto` for "canary" releases.

We're now doing normal package releases with `changesets`. The "canary"
releases still use `auto` because having a "canary" release is useful
and it would be a bit of work to set that up entirely with `changesets`.
If we find there's an easy way to do so, we can move to that approach.

The important thing here is that we're no longer relying on `auto` to
make release after we merge. Instead, we have to manually merge the
long-running PR that `changesets` will (hopefully) create/maintain for
us.

As with all changes, if we find this process not working, we can update
it to something else.