UIAPPS-87 Hoist `prettier` to top-level
This should fix the CI failures (and also be more correct).

Only by happenstance was things working before. There were no other
dependencies we had that used an incompatible version of `prettier`.
Once we added the `@changesets/cli`, it pulled in an incompatible
version of `prettier`, transitively. This caused the `lint` script to
start failing.

We address this by hoisting the `prettier` package, so it's the same
throughout the whole repo.

This is also the first changeset we're making (since the SDK changed its
dependencies).