UIAPPS-92 Fix `--directory` flag default
We want to default to the name of the example.

Before, everything would default to `starter-kit`. This is fine if
someone only ever uses the `starter-kit` example. But once they try
using one of the other examples, they get an unintuitive behavior of
downloading say the `random-dog` example to a new `starter-kit`
directory.

Instead of doing that, we use the name of the example as the default for
the directory. People can still supply the `--directory` flag
explicitly. In which case we use that as the name of the new directory.
But it's not a requirement.