Remove `pull_request_target` from `ci`
Near as we can tell, this isn't necessary anymore.

This caused an issue the other day where the `master` branch had a
changelog that wasn't formatted with to our requirements so it would
fail the `lint` step. But, the only way to get the job running on
`pull_request_target` to pass was to merge to the `master` branch.

Executing the Continuous Integration workflow on the
`pull_request_target` event seems like it might actually be a security
risk:
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/.
It seems the `pull_request_target` event should be used not for
executing code in a PR, but rather for non-code things like running
workflows or commenting on PRs:
https://docs.github.com/en/actions/learn-github-actions/events-that-trigger-workflows#pull_request_target.

Since it doesn't seem to help us make sure broken stuff doesn't get into
the `master` branch, and it seems like a potential security risk, we
remove it for now.