Change when to skip canary releases
We don't want to use `[skip<space>ci]` since that skips all workflows.

The default behavior of GitHub workflows is to skip a job if the commit
message has one of the blessed strings of text in it:
https://docs.github.com/en/actions/managing-workflow-runs/skipping-workflow-runs.
Since we were using one of those strings of text to make sure we weren't
trying to make canary releases when making actual releases, it also
meant that we were skipping all of the `pull_request`/`push` workflows.

We move over to using a different string of text (`[skip<space>canary]`)
which shouldn't match the blessed strings of text. This way, we'll still
get workflows running when making regular releases or in the long-lived
PRs.