Try to install from the global cache
Since the caching works at a global level, we try to use that first.

The way the `actions/setup-node` action works with `cache` is that it
uses a global cache. This global cache is used instead of a local cache
because `yarn` will look up dependencies correctly in it. In a local
cache, `yarn` will just assume the best and use whatever's there. That
can break when something "below" `yarn` changes (architecture, Node.js
version, etc.).

With the `--prefer-offline` flag, yarn will look in the global cache
first, then look in the remote registry for a dependency. If all goes
well, this should work out for us.