UIAPPS-88 Make all examples more consistent
We want the names to be consistent, have a version, and be private.

The name changes are two-fold:
1. It's a bit hard to figure out which package you're looking for.
    E.g. `app-starter-kit-js` was the sentiment example.
2. Some of our tooling supports working with package names as globs.

The version changes are largely for `changesets`. It seems to have an
issue with the packages not having versions even if they're not supposed
to be published.

The private changes are because we don't want to publish these packages.
It's pretty easy to accidentally publish a package, and it'd be unfun to
have to deal with that after the fact.

We put some scripts together for these constraints so we check them all
the time.