UIAPPS-88 Add scripts and run them
This gets all the formatting/linting in line for the whole repo.

There's a ton of changes here, but they're almost all lint/formatting
changes. There are some that were ignored (like console stuff). And some
of the lints required a little bit of change to address issues
(reassigning parameters, using `==` unsafely, etc.). But for the most
part, it's all consistently using the same linting/formatting.

One other thing is that we had to get all of the `webpack-dev-server`s
on the same dependency. It was causing an issue when running `yarn
build` otherwise