UIAPPS-88  Move `eslint`/`prettier` to top-level
We want to only deal with this stuff from the root of the repo.

There are a few reasons behind this change:
1. We'll check more than just workspaces.
    We have code in more than just `examples` and `packages`. We weren't
    checking any of this with each workspace having its own
    `eslint`/`prettier` setup.
2. We can be consistent with our formatting/linting.
    With each workspace having its own setup, they can get out of sync
    with each other. Rather than dealing with little changes between
    each, the root makes sure they all comply.
3. We need multiple `format`, `lint`, `lint:check`, `lint:fix` scripts.
    Before, each new package needed to get in line and have these
    scripts. Some of them weren't the same, and it also wasn't clear
    what all each needed. With it at the top-level, we can rely on one
    set of these scripts.

We had a to do quite a bit of updating to make this happen: moving
dependencies, fixing untracked errors, etc. Also, the vast majority of
this change is in formatting, but that's kind of what happens when the
root is running the show.