UIAPPS-88 Lift `typescript` to root and update it
We want a consistent version across all packages.

Before, each package could run its own `typescript` dependency. This
meant that they could (and were) out of sync. They weren't too far off,
but it was enough that even `@datadog/ui-extensions-sdk` had differences
from the others.

We now move the `typescript` dependency to the root so they all get the
same version. This also means that updating the version will have to
happen holistically, but that seems like a managable thing with the size
of the repo.

We don't remove the dependency from `examples`, since they need it when
run in isolation.

Finally, we had to update some of the `try`/`catch` behavior due to
`useUnknownInCatchVariables`:
https://www.typescriptlang.org/tsconfig#useUnknownInCatchVariables.
While that might seem annoying to deal with, the nature of JavaScript
means that if we don't handle it, we're likely going to explode at
runtime for a different issue.