UIAPPS-88 Make every example depend on latest SDK
We want everything to always be up to date.

We finally come to the meat of the problem we've been trying to address:
the examples were able to depend on anything. If we don't enforce that
the examples use the latest version, we run into a situation where code
that people are supposed to depend on is either broken or out of date.
This really bit us when we updated `@datadog/framepost` since it cause
our CI to start failling because one of the `yarn.lock` files was out of
date.

With this change, it means we'll also have to keep the versions up to
date as they change, but this should be automated by `changesets` since
it'll keep these versions up to date whenever the
`@datadog/ui-extensions-sdk` version changes. In order for that to not
cause issues, we ignore all of the example packages. This ignoring means
we only don't bump the package version (or try to release it), but we
still keep the dependencies up to date. It's confusing, but that's what
that field means.