UIAPPS-88 Fix `args` access
The newer version of the SDK types `args` as `unknown`.

Though this is more verbose, this is safer. Since `args` can be
literally anything, we don't know what we're getting and have to deal
with that somehow.

If this seems like a bit much, the resolution should be to make
receiving args from opening a side panel easier while also carrying
along type information.