UIAPPS-104 Fix UI Extension versions together
We want all `@datadog/ui-extensions-` packages to have the same version.

This is explained in more detail in the `RELEASE.md` file. The long and
short is that it should be easier on App Developers if they don't have
to manage different sets of versions, and rather can just install one
version for everything.

We add a script that checks that all the `@datadog/ui-extensions-*`
packages have a changeset if any of them do. This script works in tandem
with the change to `.changeset/config.json` where we link all
`@datadog/ui-extensions-*` packages together. It's a bit unfortunate
that we have to roll a script for this, but it's not the end of the
world. If changesets gets this behavior, we can move over to using that
instead of this two-pronged approach.

Now, we can start adding packages.