UIAPPS-104 Start a UI Extensions `react` package
We want to provide helpers for using the SDK with `react`.

We throw together a single hook for now: `useContext`. It provides the
context along with its state. We also give it a few tests to check that
it behaves the way we expect it to.

As far as the dependencies, we have two peer dependencies: `react`, and
`@datadog/ui-extensions-sdk`. Since both of these packages are being
used by our own package, and we also depend on their APIs, we want to
make sure the consumer of our package have these others installed.

This is a bit different from how dependencies normally work in the `npm`
world (but fairly normal to how dependencies work in other ecosystems).
This pattern tends to show up for similar packages to what we have here:
other `react`-based packages, plugins, or (more generally) anything
where the public API depends on a different package's API.

We also add a changeset for this new package so it'll be released.