UIAPPS-104 Use the `react` package in examples
We want to start dogfooding this package.

There's two big reasons to dogfood this:
1. It provides examples of how to actually use it.
2. It gives us feedback on whether the juice is worth the squeeze.

The first point is one that shouldn't be understated. For stuff like
this that can be kind of complex, having an example of how it's supposed
to work can be really helpful. There's a time and place for reference
documentation, and there's a time and place for example documentation.

The second point is more of an internal metric. If we look at what is
provided and don't see anything worth maintaining, we can more easily
justify dropping it. If we do see big benefits, we can more easily
justify maintaining or expanding it.

In any case, most of these examples are relatively straight-forward. The
thing that's kind of interesting is that in some cases we don't need to
use `react` hooks to get similar behavior anymore. We dropped a bunch of
`useEffect`s and even got rid of some `useState`s.

Hopefully as we expand on this `react` package, we can continue getting
this sorts of benefits.