UIAPPS-70 Setup `typedoc`
We want to provide documentation for our public packages.

The nice thing about using `typedoc` is that we'll get *something* for
everything even if we don't explicitly document stuff. It should use the
type information to construct at least the API documentation. Anything
that has jsdoc comments will also be documented and included in the
output.