UIAPPS-104 Don't bundle `react` or the SDK
We don't want to include these dependencies in our bundle.

The purpose of this package is to provide additional help atop these
other two packages. We don't want to also bring them in with us for a
few reasons:
1. The `@datadog/ui-extensions-react` package has these as peer
   dependencies.
    If we're bundling our peer dependencies, it defeats the purpose.
2. If we're bundling the `@datadog/ui-extensions-sdk`, we can run into
   issues.
   E.g. the `init` function will not work correctly if someone has both
   `@datadog/ui-extensions-react` and `@datadog/ui-extensions-sdk`
   installed and uses them both.

We tell `webpack` that these are external dependencies and how to load
them (by their same package name).