UIAPPS-104 Fix starter-kit widget
We had an infinite loop before.

The previous code was a mistake, we accidentally were calling
`setMetric` in a way that forced us to continue calling `setMetric`.

To alleviate that, we move the initial metric value up above the call to
`useState`, and use it there.

Likely, we should take this opportunity to provide better handling of
the different states the context can be in. If we're still initializing,
we can provide a different interface that if the handshake failed, and
the normal interface when things are initialized.