UIAPPS-104 Hack around the lexicographical sorting
We make a quick hack to get things working again with `yarn workspaces`.

The issue with the `yarn workspaces` command is that it takes whatever
it finds in the `"workspaces"` field and uses that in the order it's
written. It does not look at any of the dependencies of the packages
referenced and sort them topologically.

Because we finally added a package that doesn't happen to have the same
ordering lexicographically as it does topologically, all of the
`yarn workspaces` command were failing.

We first tried to use `lerna` for its topological sorting, but it both
only looks at `dependencies` (not `peerDependencies` or even
`devDependencies`) and also was ignoring all of the `examples` packages.

We considered bringing in another tool to help with this, but found out
that the `yarn workspaces` command will respect the ordering of the
`"workspaces"` field in the manifest. We do a semi-topological sort on
the packages, and let `yarn workspaces` take care of the rest.

We should circle back to this at some point, but this should be enough
to get us unblocked from being completely broken.