UIAPPS-104 Set the metric in `useEffect`
`useState` doesn't work the way it seems!

This seems to always bite my (joneshf-dd) whenever I use `useState`. It
always seems like you can pass in a value and it'll work properly, but
it really doesn't.

The value that's passed in is only used on initial render and never
again. So in this case, we were never setting the metric once the
context was initialized.

We move the setting into the `useEffect` hook so it actually works
properly.