UIAPPS-104 Convert last examples to react package
Missed this examples when doing the initial pass.

These are again showing the benefit to having this package. We don't
have to place a huge burden on App developers to understand how to work
with the `Context`. Instead, the hook can take care of most of the
intricacies and then using it is fairly straight-forward.