UIAPPS-104 Provide a facade over `useContext`
We don't want App developers to worry about the other two states.

Through discussing the previous API of `useContext`, we decided to
return an optional `Context`. The main factor in this is that when the
App is actually used in the Datadog platform, it'll never actually be
rendered unless the `Context` exists.

If there's not a `Context` available yet (because the handshake hasn't
finished), we don't render the App at all. If there's a failure in the
handshake, we also don't render the App at all. From the perspective of
the App developer, it's a bit of a mis-step to attempt to render
something if there's no `Context`. Rather than supporting that pattern,
we combine all the non-existent states together into `undefined`.

There's still more we can do here, like not rendering the component at
all if it doesn't have a context. But, we hold off on any of that until
later.