Cache `node_modules` explicitly
We factor the Node.js version into the cache key.

The main reason to not cache `node_modules` seems to be that it can
break when the Node.js version changes:
https://github.com/actions/cache/blob/main/examples.md#node---npm. If we
use the `.node-version` file as part of the cache key, it should bust
the cache whenever we change the Node.js version.

Hopefully, this will mean that we shouldn't have issues with caching
`node_modules`. While also meaning we can get some benefits from caching.