UIAPPS-141 Move `RequireKeys` into `types` module
We only use it in there, so we move it in there.

This is technically a breaking change since it's exported from
`utils/utils` and we bundle that up when we ship the package. Anyone can
access this by importing `@datadog/ui-extensions-sdk/dist/utils/utils`.
It's not intentional that it was available before, but it was available
before.

This change lets us remove one of the dependency cycles between modules.
In particular, `types` was importing `utils/utils` and vice versa.

We add a changeset so we'll get a new release once this goes out.