UIAPPS-141 Break cycle for `Logger`
We introduce an interface and code to it.

Instead of having both `DDClient` depend on `Logger` and `Logger` depend
on `DDClient`, we pull up an interface that `Logger` can depend on and
code to that. We make sure `DDClient` implements that interface and move
on from there.

This is the general pattern we're going to try and follow for these
dependency cycle breaks.