UIAPPS-141 Introduce event and request interfaces
We want to break the cycles from needing to use `framepost` stuff.

We introduce two new interfaces for dealing with events and requests. We
make sure the `DDClient` implements them, and depend on only them
wherever possible. There's still quite a few places that still need the
whole `DDClient`, but this should be the majority of uses.

With those hard dependencies broken, we can actually make the
`framePostClient` property private to `DDClient` only. Consumers can
still do basically everything they could before, except they don't have
direct explicit access to `framepost`'s public API.

There's one part to this that might need further explanation. In the
`DDFeatureClient`, we were calling
`this.client.framePostClient.handshake` explicitly so it would throw
an error if it wasn't setup properly. Given that that code next calls
its own `isEnabled` method, which calls `this.client.getContext`, which
eventually calls the same handshake method, it should have effectivelys
the same behavior. Namely, it'll throw if the handshake failed. We call
this out explicitly, because it's not readily clear why we were able to
remove that code without any other changes.