UIAPPS-141 Add log interface and break last cycles
We make all the dependencies flow in one direction.

The `LoggerClient` wasn't really the keystone to everything, it was just
the last brick that was put in place. Now that it's there, we can break
all the rest of the cycles by depending on the interfaces rather than
the specific implementation of `DDClient`.