Fix `yarn build` for geomap App
It was using a different version of `react-scripts`.

Since `react-scripts` is very opinionated about which versions get used,
and we're in a monorepo, they have to all line up across all packages.
This new example was using `4.0.1`, while the others were using `4.0.3`.
We update this to match the rest of the packages.

Co-authored-by: tdimnet <thomas.dimnet@datadoghq.com>