Start working on a Vue example
We get the brunt of the setup, though there's more to do after this.

This is enough to have a widget show up in Datadog, though it doesn't do
a whole lot. The point of this example isn't for it to be a kitchen sink
like the starter-kit. It's to show a minimal example in Vue. There are a
couple of things to be aware of though:
1. This doesn't fit with our package requirements.
    We'll have to update some of the infrastructure here to make sure it
    passes the requirements and stays consistent.
2. This is not using the local SDK.
    This goes to the previous point, but we got stuck trying to figure
    out why it wouldn't work. Since we know there are Apps out there
    written in Vue, it seems like there's an issue with this example
    being inside a yarn workspace. For now, we decided to target an
    older version of the SDK. We'll circle back to try and figure it
    out.

Co-authored-by: Alec Barrett <alec.barrett@datadoghq.com>