Fix consistency issues
We want to fall in line with the other examples.

The two changes that probably aren't straight-forward are the
`vue.config.js` and the `"nohoist"` in the `package.json`.

For `vue.config.js`, we need to override the webpack settings for
symlinks within this monorepo. Turns out there's a known issue with
webpack and symlinks:
https://webpack.js.org/configuration/resolve/#resolvesymlinks. There's a
workaround for Vue we can use:
https://github.com/vuejs/vue-cli/issues/1494#issuecomment-498144990. So,
we use it. This allows us to reference the local version of the SDK and
use it. This also means we can be consistent with the other examples.

For `"nohoist"`, we ran into another conflict with the dependencies
being hoisted to the top-level, but it seemed impossible to fix by just
being more specific in the workspaces. The backstory here is that we've
consistently run into this issue before where some dependency is hoisted
to the top-level and `react-scripts` blows up because it's not the
version it was expecting.

For the most part, we end up keeping similar dependencies in each
different package. Because they all use different versions, none of them
are hoisted. But now, there's multiple dependencies in the
`datadog-app-example-vue-custom-widget` package that depend on the same
version of `webpack`. Since multiple dependencies depend on the same
version, yarn decides to hoist it to the top-level, which causes
`react-scripts` to explode.

We work around that by using the `"nohoist"` option:
https://classic.yarnpkg.com/blog/2018/02/15/nohoist/. We probably
should've used this from the start, but here we are. Hopefully, this
will make these sorts of issues less prevalent in the future.

The rest of these changes are just making things consistent (read: pass
CI).