Fix error with Controller route
Apparently, leaving off the `component` works, but logs an error.

If we leave off the `component`, it'll still be able to work and it'll
still work as an App. But, it will print a warning to the console and
also throw an error.

To work around that, we add a simple Controller view that works like the
starter-kit's controller does.