Hook up the metric option
We make this custom widget behave more like the starter-kit.

We follow much the same pattern, grab the initial value, setup an event
listener, render the metric. We also clean up the listener when we're
done. This is based off of the lifecycle diagram:
https://v3.vuejs.org/guide/instance.html#lifecycle-diagram. If there's a
more idiomatic way to do this, we can switch to that instead.