Track the manifest
We can use this to import into the Developer Platform.

Probably the most important part of tracking this is that we can show
what the `source` for the widget needs to be with the `vue-router`. It
has to be the hash-route: `#/widget`. If we attempt to leave off the
hash, it will always go to the controller.