UIAPPS-93 Add a `--install` option
This is useful to avoid having to install everything.

This is especially useful when trying to test this package. If we have
to do a full install each time, we're opening up ourselves to flakiness
and also slowness. We can eschew that whole thing by allowing ourselves
to turn off the install behavior when we want.