UIAPPS-93 Add a `--features` option
This option alllows initializing with specific features.

To get to the point where App developers can create an App with just
what they need, we add this option that allows specifying only the
features they want.

This option currently doesn't do a whole lot, except make sure it's not
used with the `--example` option. We'll flesh it out more as we move
along.

We also setup an acceptance test for making sure this work works as we
expect. We don't really need to check everything, but we at least want
to check a few things that are important.