UIAPPS-151 Fix updating an App
We require the `id` for the App when updating it.

We also update (and document), the uses. The documentation is there
because it's a little hard to understand what the `appId` is for in
plain JS.

Taking a step back, it doesn't really make sense to pass in the
`endpoint` or `method` and force the caller to also remember to pass in
the App id when it's appropriate. Rather, that function should figure it
out itself based on the App id.